backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
PyJWT==2.9.0
python-dotenv==1.0.1
django-filter==24.3
Pillow==10.4.0
